refactor save logic
